socket.io-send-receive/requirements.txt
python-socketio==5.14.1
aiohttp==3.13.0